requirements.txt
Flask==3.0.0
flask-login==0.6.3
werkzeug==3.0.1
openai==2.14.0
python-dotenv==1.0.0
markdown==3.5.1
reportlab==4.0.7